Add support for specifying a target path parameter, which allows a request parameter to be configured with a destination to redirect to after authenticating.
This allows a URL such as /connect/facebook?_target_path=/some/place to
link directly to an authentication provider, bypassing the login screen,
while still redirecting to a choosable location.

Redirection works by setting the appropriate target path value in the
session in the same way that it is done
Symfony\Component\Security\Http\Firewall\ExceptionListener when an
unauthenticated user requests a URL which is behind a firewall.