Fix revoke token for github.
https://developer.github.com/v3/oauth_authorizations/#revoke-an-authorization-for-an-application

The old code did not work for me. Judging from the commit that added it i
assume it was part of a copy+paste job. According to the documentatio from
github to revoke a token you need to send a DELETE request. And its a one
step process.

Works for me.