Create entity for State parameter
Currently the bundle only supports passing a CSRF token in the state
parameter. We should be able to pass more than just that in the state
parameter, such as any other type of state we want to round-trip back to
the application.
This value object supports passing a single value for CSRF tokens OR
state, or several values in an encoded array. It does the conversion to
and from encoded array/string.