Make state parameters configurable
This commit enables the user to add a `state` key to a resource owner's
configuration, which can be populated with a single string or an
associative array. That configuration will then be added to the resource
owner upon construction. All the parameters will be encoded and sent
back as a json/base64/url encoded string.
CSRF token support is still enabled and backwards compatible. If only
the CSRF token is set, that will be the only thing in the state after it
makes the round trip (plain-text, not encoded).