Fix controller syntax in routing files
As reported in #1522, this bundle routing is currently broken.

Symfony supports two syntax for the _controller key in routing files:
* `FQCN::action`
* or `service_name:action`

It appears both syntax have been mixed up together: the first syntax
only works when the service name is a FQCN, whereas the controller
service is currently named using a custom string (non-FQCN).