Re-add state support (#1640)
* Support for multiple OAuth2 state parameters

Currently the bundle only supports passing a CSRF token in the state
parameter. We should be able to pass more than just that in the state
parameter, such as any other type of state we want to round-trip back to
the application.

This value object supports passing a single value for CSRF tokens OR
state, or several values in an encoded array. It does the conversion to
and from encoded array/string.

* Added state support from #1456 and #1549

* Fixed styling issues

* Added support for passing in state parameters in login urls

* Fixed styling issues

* Fixed review suggestions

* Review fixes

* Added string of strings to StateInterface::getAll() as return value

* Added null support for state

* Fixed styling

Co-authored-by: paulinevos <pvos88@gmail.com>
Co-authored-by: Balazs Csaba <balazscsaba2006@gmail.com>