Fixed a double credentials error
When the "use_authorization_to_get_token" option is on (which it by default is), credentials are sent twice if they are also present in the parameters.

Additionally, client ID and secret are always overwritten in the parameters if this option is turned off, even in cases where the secret is generated and already added to the parameters (for example in `AppleResourceOwner` and `InstagramResourceOwner`).